requirements.txt
discord.py>=2.2.2
openai>=0.27.2
requests>=2.28.2
wikipedia>=1.4.0
textblob>=0.16.0
tinydb>=4.5.2
pytz>=2023.3
jsonpickle>=3.0.1
nltk>=3.8.1
tinydb_serialization>=2.1.0
retry>=0.9.2
sortedcollections>=2.1.0